[@santihsdev] Updated set up to v11 folke/lazy.vim

+++ copy.sh
@@ -1,5 +1,25 @@
-sudo cp -r ~/.config/lvim ./lvimv2.0
+if [ -d ~/.config/lvim ]; then
+    sudo cp -r ~/.config/lvim ./lvimv2.0
+else
+    echo "~/.config/lvim does not exist, skipping copy"
+fi
+
 sudo rm -rf ./lazyvim-santihs
-sudo cp -r ~/.config/nvim ./lazyvim-santihs
-cp ~/.ideavimrc ./
-cp ~/.zshrc ./
+
+if [ -d ~/.config/nvim ]; then
+    sudo cp -r ~/.config/nvim ./lazyvim-santihs
+else
+    echo "~/.config/nvim does not exist, skipping copy"
+fi
+
+if [ -f ~/.ideavimrc ]; then
+    cp ~/.ideavimrc ./
+else
+    echo "~/.ideavimrc does not exist, skipping copy"
+fi
+
+if [ -f ~/.zshrc ]; then
+    cp ~/.zshrc ./
+else
+    echo "~/.zshrc does not exist, skipping copy"
+fi
